django/requirements.txt
asgiref==3.11.1
Django==6.0.2
djangorestframework==3.16.1
sqlparse==0.5.5

# AI/LLM dependencies
langchain-google-genai>=2.0.0
langchain-core>=0.3.0
jsonschema>=4.0.0
python-dotenv>=1.0.0
